[CHORE] #231 바인딩 어댑터 패키지 위치 통일하기
