+++ app/src/main/res/layout/activity_end_run.xml
@@ -65,9 +65,9 @@
 
         <androidx.appcompat.widget.AppCompatImageView
             android:id="@+id/iv_endRun_capture"
+            loadEndRunImage="@{model.captureUri.toString()}"
             android:layout_width="match_parent"
             android:layout_height="362dp"
-            app:endRunImage="@{model.captureUri.toString()}"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@id/storage_toolbar"
