+++ app/src/main/res/layout/item_storage_my_draw.xml
@@ -40,9 +40,9 @@
 
                 <androidx.appcompat.widget.AppCompatImageView
                     android:id="@+id/iv_my_page_upload_course"
+                    loadStorageImage="@{storageItem.image}"
                     android:layout_width="match_parent"
                     android:layout_height="match_parent"
-                    app:storageImage="@{storageItem.image}"
                     android:scaleType="centerCrop" />
 
             </androidx.cardview.widget.CardView>
@@ -54,12 +54,12 @@
                 android:elevation="3dp"
                 android:outlineProvider="none"
                 android:visibility="invisible"
-                app:srcCompat="@drawable/check_box_selector"
                 app:scale_height="20"
                 app:scale_left="8"
                 app:scale_top="8"
                 app:scale_width="20"
-                />
+                app:srcCompat="@drawable/check_box_selector" />
+
             <androidx.appcompat.widget.AppCompatImageView
                 android:id="@+id/iv_my_page_upload_course_select_background"
                 android:layout_width="wrap_content"
